Fix Ruff formatting, Mypy typing mismatches and add beartype to requirements

+++ diffusionkit/utils.py
@@ -46,7 +46,7 @@ def _load_weights(module: nn.Module, path: str) -> None:
 
 def bytes2gigabytes(n: int) -> int:
     """Convert bytes to gigabytes"""
-    return n / 1024**3
+    return int(n / 1024**3)
 
 
 def image_psnr(reference: Image, proxy: Image) -> float:
